迁移：原始动态数组 → std::vector，修复 8 个 bug
核心变更：
- 所有 new[]/delete[] 替换为 std::vector / std::vector<std::vector<double>>
- 删除 ~11 个手动析构函数（RAII 自动管理）
- Ax_b::exchange_row 用 std::swap 实现 O(1) 行交换
- filelog 改为接受 const vector&（避免拷贝）
- 范数函数统一为 vector 版本，删除原始指针重载
- add_point() 从 34 行手动重分配 → 6 行 push_back/resize

Bug 修复：
- cube_spline case 2 的 M_solve[-1] 越界 → M_solve[i]
- exchange_diag_no_0 负数索引检查顺序 + 外层死循环
- conjugate_gradient 忽略用户设置的 maxcounter
- 8 个类 else ClassName() 创建临时对象而非委托构造
- Householder(bool) 的 print_flag 未初始化
- 缺少 #include <algorithm>
- int_diff.cpp 两处 double* e 内存泄漏

其他：
- 消除全部 -Wsign-compare 和 -Wunused-parameter 警告
- 更新 CLAUDE.md（内存管理、函数约定、已知问题）
- 更新 .gitignore

+++ src/eigen_val_vec.cpp
@@ -183,7 +183,7 @@ void Power_method::out_result() {
 #endif
 	cout << "xi =" << endl;
 	fl << "\nxi =";
-	for (int i = 0; i < z.size(); i++) {
+	for (size_t i = 0; i < z.size(); i++) {
 		cout << "  " << z[i] << endl;
 		fl << "\n  " << z[i];
 	}
@@ -294,7 +294,7 @@ void Inverse_power::out_result() {
 #endif
 	cout << "xi =" << endl;
 	fl << "\nxi =";
-	for (int i = 0; i < z.size(); i++) {
+	for (size_t i = 0; i < z.size(); i++) {
 		cout << "  " << z[i] << endl;
 		fl << "\n  " << z[i];
 	}
